delete Modbus libaray and test it and made change

+++ H1DR1/H1DR1.h
@@ -132,7 +132,7 @@ extern UART_HandleTypeDef huart5;
 extern UART_HandleTypeDef huart6;
 
 /* Define UART Init prototypes */
-Module_Status MB_PORT_Init(uint16_t BaudRate, uint32_t DataBitsN, uint32_t ParityBit, uint32_t StopBitN);
+Module_Status MB_PORT_Init(uint32_t BaudRate, uint32_t DataBitsN, uint32_t ParityBit, uint32_t StopBitN);
 extern void MX_USART1_UART_Init(void);
 extern void MX_USART2_UART_Init(void);
 extern void MX_USART3_UART_Init(void);
